Numericaly Instability Issue Solved!
I finally identified the numerical instability issue.  The problem was that the $C^{\infty}$-bump function (classical Schwarz-type test function) was numerically unstable even if it is infinitely-differentiable...

So instead the bump function (x-σ)^2(x+σ)^2I_{x< σ} is stable and does the required job even if it is only C^1! Wohooo poyoo